requirements.txt
altgraph==0.17.5
arabic-reshaper==3.0.0
AwesomeTkinter==2021.11.8
certifi==2025.11.12
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
evdev==1.9.2
idna==3.11
packaging==25.0
pillow==12.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pynput==1.8.1
pyperclip==1.11.0
pystray==0.19.5
python-bidi==0.6.7
python-xlib==0.33
python_appimage==1.4.5
requests==2.32.5
setuptools==80.9.0
six==1.17.0
urllib3==2.6.2
